add `Gui`, `Gui, Show`, `Gui, Add`, `text` gui control, control positioning and internal`parse_word_options`
positioning accepts `x`, `xpn`, `xmn`, `xsn`, `w`, but not `xn`, `wpn` or `r`. Same thing for `x`/`h`. (`n` refers to an optional number, optionally prefixed with `+` or `-`) `section` also added.
In other words, gui control positioning is already complete except for those options that rely on the width/height of the previous control, and also `r` (rows) not.
Padding is currently fixed 7px, should be based on font size of first control instead, this is TBD when fonts are added