add AT/SPI (control support)
this is the basis for all sorts of control- and text-based commands.

there may be different prerequisites necessary for atspi to work. These can barely be automated and if we attempted to do so, it would be too intrusive to the user, so instead maybe as a user ahk script some day. Instead, when data cannot be accessed, a gigantic popup shows to the user, explaining the options they have. This is probably not the best solution, but it is necessary to somehow inform the end user here because the situation may differ from system to system and cannot be reliably planned for by the script writer. Best add a hidden #NoAtspiWarn directive some day.

I tested this (commands will follow the next days) with
- many different gtk and qt programs
- firefox / librewolf
- chromium
- electron apps: signal, vscodium
- libreoffice
- a java swing gui
and while accessibility support for all of these is only half baked and buggy annd all needed some tweaks, the resulting commit code should be able to nicely target each of those platforms.

There is no unique for elements in atspi, so it was necessary to artificially craft classNNs. They are based on rule + indexes path.

closes #3