fix star hotkeys matching shift modifier too
example: `*a::` wasn't triggered by <kbd>shift</kbd>+<kbd>a</kbd> but now is
Internals slightly changed with a small breaking change: Hotkey with shift keys need to be declared explicitly because the uppercase version is transformed to lowercase. So it isn't enough anymore to do e.g. `A::` - you'll have to do `+A::` or `+a::` now.