fix Hotkey/Send combination on Solus Linux
key / keyboard grabbing was broken because typing a hotkey still flashes the current window which leads to the focus being taken away and set to the root window 0 instead. Our current mechanism then ungrabbed and re-grabbed on the root window every time. This works fine in theory, but leads to race conditions when Sends are happening at the same time.
Instead of solving said race condition, I simply prevented grabbing on the root window now.  It seems that the hotkeys are still grabbed, even though the last-grabbed window is not in focus anymore. That's probably because the root window is magical, but who knows. It seems to work fine on all distros now, and the missing root grabs will even improve performance on Solus.

Fixes #51